:recycle: refactor: String -> BoardCategoryType [#32]
