+++ dongamboard/app/src/main/java/com/study/dongamboard/model/request/PostRequest.kt
@@ -2,6 +2,7 @@ package com.study.dongamboard.model.request
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
+import com.study.dongamboard.type.BoardCategoryType
 import java.io.Serializable
 
 @JsonClass(generateAdapter = true)
@@ -13,5 +14,5 @@ data class PostRequest (
     val content: String,
 
     @field:Json(name = "category")
-    val category: String,
+    val category: BoardCategoryType,
 ) : Serializable
