+++ dongamboard/app/src/main/java/com/study/dongamboard/type/BoardCategoryType.kt
@@ -1,0 +1,6 @@
+package com.study.dongamboard.type
+
+enum class BoardCategoryType {
+    자유게시판,
+    장터게시판;
+}
